requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
line-bot-sdk>=3.11,<4.0
anthropic>=0.40,<1.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
